+++ docs/common/error-utilities.ts
@@ -474,4 +474,11 @@ const RENAMED_PAGES: Record<string, string> = {
 
   // After consolidating the "Internal distribution" information
   '/guides/sharing-preview-releases/': '/build/internal-distribution/',
+
+  // After moving from eas-workflows to eas/workflows
+  '/eas-workflows/get-started/': '/eas/workflows/get-started/',
+  '/eas-workflows/triggers/': '/eas/workflows/syntax/#on',
+  '/eas-workflows/jobs/': '/eas/workflows/syntax/#jobs',
+  '/eas-workflows/control-flow/': '/eas/workflows/syntax/#control-flow',
+  '/eas-workflows/variables/': '/eas/workflows/syntax/#jobsjob_idoutputs',
 };