+++ packages/expo-dev-launcher/android/src/rn74/debug/expo/modules/devlauncher/react/DevLauncherDevSupportManagerSwapper.kt
@@ -16,8 +16,6 @@ import expo.modules.devlauncher.helpers.getProtectedFieldValue
 import expo.modules.devlauncher.helpers.setProtectedDeclaredField
 import expo.modules.devlauncher.koin.DevLauncherKoinComponent
 import expo.modules.devlauncher.launcher.DevLauncherControllerInterface
-import expo.modules.devlauncher.react.DevLauncherBridgeDevSupportManager
-import expo.modules.devlauncher.react.DevLauncherBridgelessDevSupportManager
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.launch
 import org.koin.core.component.inject