+++ packages/expo-ui/ios/SectionView.swift
@@ -12,12 +12,12 @@ struct SectionView: ExpoSwiftUI.View {
   @EnvironmentObject var props: SectionProps
 
   var body: some View {
-    var form = Form {
+    let form = Form {
       Section(header: Text(props.title ?? "")) {
         Children().padding(EdgeInsets(top: 0, leading: 0, bottom: props.heightOffset, trailing: 0))
       }
     }
-    
+
     if #available(iOS 16.0, tvOS 16.0, *) {
       form.scrollDisabled(true)
     } else {